Fix new-folder ipc listener
